requirements.txt
# Collecte
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1

# Traitement
pandas==2.2.0
langdetect==1.0.9

# ML
transformers==4.38.0
torch==2.2.0

# Dashboard
streamlit==1.32.0
plotly==5.20.0

# Utils
python-dotenv==1.0.1
